+++ packages/amplify-graphql-api-construct-tests/src/__tests__/backends/base-cdk/app.ts
@@ -2,7 +2,7 @@
 import 'source-map-support/register';
 import { App, Stack, Duration } from 'aws-cdk-lib';
 // @ts-ignore
-import { AmplifyGraphqlApi, AmplifyGraphqlSchema } from '@aws-amplify/graphql-construct-alpha';
+import { AmplifyGraphqlApi, AmplifyGraphqlDefinition } from '@aws-amplify/graphql-api-construct';
 
 const packageJson = require('../package.json');
 
@@ -13,12 +13,12 @@ const stack = new Stack(app, packageJson.name.replace(/_/g, '-'), {
 
 new AmplifyGraphqlApi(stack, 'GraphqlApi', {
   apiName: 'MyGraphQLApi',
-  schema: AmplifyGraphqlSchema.fromString(/* GraphQL */ `
+  definition: AmplifyGraphqlDefinition.fromString(/* GraphQL */ `
     type Todo @model @auth(rules: [{ allow: public }]) {
       description: String!
     }
   `),
-  authorizationConfig: {
+  authorizationModes: {
     apiKeyConfig: { expires: Duration.days(7) },
   },
 });