+++ packages/amplify-graphql-api-construct/src/index.ts
@@ -4,9 +4,8 @@ export type {
   OIDCAuthorizationConfig,
   ApiKeyAuthorizationConfig,
   LambdaAuthorizationConfig,
-  AuthorizationConfig,
-  PartialSchemaTranslationBehavior,
-  AmplifyApiSchemaPreprocessorOutput,
+  AuthorizationModes,
+  PartialTranslationBehavior,
   AmplifyGraphqlApiProps,
   AmplifyGraphqlApiResources,
   AmplifyGraphqlApiCfnResources,
@@ -23,13 +22,11 @@ export type {
   AutomergeConflictResolutionStrategy,
   ConflictResolutionStrategyBase,
   ConflictResolutionStrategy,
-  SchemaTranslationBehavior,
-  IAmplifyGraphqlSchema,
+  TranslationBehavior,
+  IAmplifyGraphqlDefinition,
   IBackendOutputStorageStrategy,
-  BackendOutputEntry,
+  IBackendOutputEntry,
+  AddFunctionProps,
 } from './types';
 export { AmplifyGraphqlApi } from './amplify-graphql-api';
-export { AmplifyGraphqlSchema } from './amplify-graphql-schema';
-
-// remove these exports when provided by cli
-export { GraphqlOutput, versionedGraphqlOutputSchema } from './graphql-output';
+export { AmplifyGraphqlDefinition } from './amplify-graphql-definition';