+++ packages/amplify-graphql-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer@1.2.0...@aws-amplify/graphql-transformer@1.2.1) (2023-10-02)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer
+
 # [1.2.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer@1.1.4...@aws-amplify/graphql-transformer@1.2.0) (2023-09-20)
 
 ### Features